requirements.txt
six
attrs
